+++ lib/src/extensions/benefits.dart
@@ -1,5 +1,4 @@
-import 'package:moment_dart/src/localization.dart';
-import 'package:moment_dart/src/moment.dart';
+import 'package:moment_dart/moment_dart.dart';
 
 extension MomentBenefits on DateTime {
   /// Returns if [year] is leap year.
@@ -151,10 +150,12 @@ extension MomentBenefits on DateTime {
   }
 
   DateTime clone() {
-    final value = DateTime(
+    if (isUtc) {
+      return DateTime.utc(
+          year, month, day, hour, minute, second, millisecond, microsecond);
+    }
+    return DateTime(
         year, month, day, hour, minute, second, millisecond, microsecond);
-
-    return isUtc ? value.toUtc() : value;
   }
 
   Moment toMoment({MomentLocalization? localization}) =>
@@ -182,29 +183,60 @@ extension MomentBenefitsPlus on Moment {
 
   Moment clone() {
     return Moment(
-        DateTime(
-            year, month, day, hour, minute, second, millisecond, microsecond),
-        localization: localization);
+      forcedSuperType.clone(),
+      localization: localization,
+    );
   }
 }
 
 extension DurationExtra on Duration {
   static const int daysPerWeek = 7;
 
+  /// Only returns whole weeks. Here, a week equals [daysPerWeek] days
   int get inWeeks => inDays ~/ daysPerWeek;
 
   static const int daysPerMonth = 30;
+
+  /// Precise number for nerds
   static const double daysPerMonthPrecise = 30.4368499;
 
+  /// Only returns whole months. Here, a month equals [daysPerMonth] days
   int get inMonths => inDays ~/ daysPerMonth;
 
   static const int daysPerYear = 365;
+
+  /// Precise number for nerds
   static const double daysPerYearPrecise = 365.242199;
 
+  /// Only returns whole years
   int get inYears => inDays ~/ daysPerYear;
 
   static const int weeksPerYear = 52;
+
+  /// Precise number for nerds
   static const double weeksPerYearPrecise = 52.177457;
 
   static const int monthsPerYear = 12;
+
+  String toDurationString(
+    MomentLocalization localization, {
+    bool round = true,
+    bool omitZeros = true,
+    bool includeWeeks = false,
+    UnitStringForm form = UnitStringForm.full,
+    String? delimiter,
+    DurationFormat format = DurationFormat.auto,
+    bool dropPrefixOrSuffix = false,
+  }) {
+    return localization.duration(
+      this,
+      round: round,
+      omitZeros: omitZeros,
+      includeWeeks: includeWeeks,
+      form: form,
+      delimiter: delimiter,
+      format: format,
+      dropPrefixOrSuffix: dropPrefixOrSuffix,
+    );
+  }
 }